+++ src/main/java/blackjack/domain/card/CardSymbol.java
@@ -1,8 +1,18 @@
 package blackjack.domain.card;
 
 public enum CardSymbol {
-    SPADE,
-    DIAMOND,
-    HEART,
-    CLUB;
+    SPADE("스페이드"),
+    DIAMOND("다이아몬드"),
+    HEART("하트"),
+    CLUB("클로버");
+
+    private final String value;
+
+    CardSymbol(String value) {
+        this.value = value;
+    }
+
+    public String get() {
+        return value;
+    }
 }
