+++ src/main/java/blackjack/domain/gamer/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.gamer;
 
 public class Player extends Gamer {
     private static final String NAME_EMPTY_ERROR = "공백이 아닌 플레이어를 입력해 주세요.";
