+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -1,6 +1,7 @@
 package blackjack.domain;
 
 import blackjack.domain.card.CardPicker;
+import blackjack.domain.gamer.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
